Moved component stories to `.storybook` folder
