+++ .storybook/main.ts
@@ -5,8 +5,6 @@ const config: StorybookConfig = {
   "stories": [
     "./stories/**/*.mdx",
     "./stories/**/*.stories.@(js|jsx|mjs|ts|tsx)",
-    "../src/**/*.mdx",
-    "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"
   ],
   "addons": ["@storybook/addon-docs"],
   "framework": {
