+++ .storybook/stories/Components/Separator/Separator.mdx
@@ -1,6 +1,7 @@
 import { Meta, ArgTypes } from '@storybook/addon-docs/blocks';
 import * as Separator from './Separator.stories';
-import * as CssVars from '../../../.storybook/stories/CssVarsExample/CssVarsExample.stories.ts';
+import * as CssVars from '../../Examples/CssVarsExample/CssVarsExample.stories';
+
 
 <Meta of={Separator} />
