+++ .storybook/stories/Components/SubMenu/SubMenu.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { fn } from 'storybook/test';
 
-import { SubMenu } from './SubMenu';
+import { SubMenu } from '../../../../src/components/SubMenu/SubMenu';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
