+++ .storybook/stories/Types/MenuItemProps.mdx
@@ -1,5 +1,5 @@
 import { Meta, ArgTypes } from '@storybook/addon-docs/blocks';
-import * as MenuItem from '../../../src/components/MenuItem/MenuItem.stories.ts';
+import * as MenuItem from '../Components/MenuItem/MenuItem.stories.ts';
 
 <Meta title='Types/MenuItemProps' />
