+++ .storybook/stories/Types/Win32MenuCSSVars.mdx
@@ -1,5 +1,5 @@
 import { Meta, ArgTypes } from '@storybook/addon-docs/blocks';
-import * as meta from '../../stories/CssVarsExample/CssVarsExample.stories.ts';
+import * as meta from '../Examples/CssVarsExample/CssVarsExample.stories.ts';
 
 <Meta title='Types/Win32MenuCSSVars' />
