+++ .storybook/stories/Types/Win32MenuStyleProps.mdx
@@ -1,5 +1,5 @@
 import { Meta, ArgTypes } from '@storybook/addon-docs/blocks';
-import * as meta from '../../stories/FullMenuBar/FullMenuBar.stories.ts';
+import * as meta from '../Examples/FullMenuBar/FullMenuBar.stories.ts';
 
 <Meta title='Types/Win32MenuStyleProps' />
